+++ backend/src/api/blog-page/routes/blog-page.ts
@@ -1,0 +1,6 @@
+/**
+ * Blog Page Routes
+ */
+import { factories } from '@strapi/strapi';
+
+export default factories.createCoreRouter('api::blog-page.blog-page');